# Conflicts:
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Air/PF_FlyingPig.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Cage/PF_Cage.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Dragon/PF_EnemyTier1.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Dragon/PF_EnemyTier3.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Dragon/PF_EnemyTier4.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Goblin/PF_Guardian.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Junk/PF_BadJunkBone.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Junk/PF_BadJunkFrog.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Junk/PF_BadJunkMagicBottle.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Junk/PF_RouletteFrog.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Seasonal/PF_PumpkinBucket_0.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Seasonal/PF_PumpkinBucket_1.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Seasonal/PF_PumpkinBucket_2.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Seasonal/PF_Seasonal.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Seasonal/PF_XmasCandy_1.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Seasonal/PF_XmasCandy_2.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Water/PF_Crocodile.prefab
#	Assets/Art/3D/Gameplay/Entities/Prefabs/Water/PF_Crocodile_Evade.prefab
#	Assets/Game/Gameplay/Entities/AI/View/ViewControl.cs
#	Assets/Game/Gameplay/Entities/Cage/CageBehaviour.cs


Former-commit-id: 31d53e4e716e57675565cbd0125a4a15d0834d02

+++ Assets/Game/Gameplay/Entities/IEntity.cs
@@ -60,7 +60,7 @@ abstract public class IEntity : ISpawnable {
 	protected float m_health;
 	public float health { get { return m_health; } set { m_health = value; } }
 
-	public virtual int score { get { return 0; } }
+	public virtual float score { get { return 0; } }
 
 	[SerializeField]
 	public ISpawnable[] m_otherSpawnables;
